requirements.txt
Flask
python-dotenv
requests
azure-cognitiveservices-speech
azure-ai-textanalytics
azure-core
azure-monitor-opentelemetry
opentelemetry-api
opentelemetry-sdk
gunicorn